refactor: extract AmountAndFiatText
